# crud-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/admin/stores": {
            "get": {
                "tags": [
                    "admin/store"
                ],
                "operationId": "daacb93aaa654b4fdc47881f94600ed6",
                "responses": {
                    "200": {
                        "description": "Lista todas las tiendas."
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "admin/store"
                ],
                "operationId": "5e3f599d3d84279ebc59e8f1be7715e7",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "location",
                                    "latitude",
                                    "longitude",
                                    "icon",
                                    "banner",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "GoodMeal market"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Av. Francisco Bilbao 2429"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Av nueva los leones 50, Providencia, Santiago, (Caracol los leones) Zona de delivery Store central"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "example": "-6.487795"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "example": "-76.370540"
                                    },
                                    "icon": {
                                        "type": "file",
                                        "example": ""
                                    },
                                    "banner": {
                                        "type": "file",
                                        "example": ""
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "08:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "15:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Datos de la tienda creada"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/stores/{id}": {
            "get": {
                "tags": [
                    "admin/store"
                ],
                "description": "Envia como parámetro el ID de una tienda y le retorna sus datos. El usuario autorizado tiene que ser el creador de la tienda",
                "operationId": "8a6fa02537b5ff997abb4a174fb9a3d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra datos de la tienda"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "admin/store"
                ],
                "operationId": "727c234b1c3ce86737c75e6b7f75d427",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo nombre"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Otra dirección"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Otra locación"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "example": "-5.487795"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "example": "-86.370540"
                                    },
                                    "icon": {
                                        "type": "file",
                                        "example": ""
                                    },
                                    "banner": {
                                        "type": "file",
                                        "example": ""
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "09:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "18:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Not content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "admin/store"
                ],
                "description": "Envia como parámetro el ID de una tienda y se eliminan los datos. El usuario autorizado tiene que ser el creador de la tienda",
                "operationId": "55a75c1bd29ccdb6875aa6192870540d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Not Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "description": "Envia sus credenciales de acceso y obtendrá un token de autorización. Use el token para endpoints protedigos enviando en la cabecera de la petición: 'Authorization: Bearer YOUR_TOKEN'",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "description": "The Token Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "mflores@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "TOKEN"
                    },
                    "401": {
                        "description": "Credenciales incorrectas."
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "responses": {
                    "200": {
                        "description": "Lista todas las órdenes."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "description": "Envia como parámetro el ID de una órden y le retorna sus datos",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la órden a consultar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra datos de la órden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/stores": {
            "get": {
                "tags": [
                    "stores"
                ],
                "operationId": "789854b8f9730d47f1ba52ca5e610238",
                "responses": {
                    "200": {
                        "description": "Lista todas las tiendas con stock."
                    }
                }
            }
        },
        "/api/stores/{id}": {
            "get": {
                "tags": [
                    "stores"
                ],
                "description": "Envia como parámetro el ID de una tienda y le retorna sus datos",
                "operationId": "cd24cde4c43f6c4aeb2ed9725fca9624",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra datos de la tienda"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer YOUR_TOKEN)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}